+++ crates/config/src/widgets/workspace.rs
@@ -1,7 +1,8 @@
 use crate::common::CommonConfig;
-use crate::shared::Curve;
+use crate::shared::{
+    color_translate, option_color_translate, schema_color, schema_optional_color, CommonSize, Curve,
+};
 
-use super::common::{self, CommonSize};
 use cosmic_text::Color;
 use schemars::Schema;
 use schemars::{json_schema, JsonSchema};
@@ -38,20 +39,20 @@ pub struct WorkspaceConfig {
     pub pop_duration: u64,
 
     #[serde(default = "dt_default_color")]
-    #[serde(deserialize_with = "common::color_translate")]
-    #[schemars(schema_with = "common::schema_color")]
+    #[serde(deserialize_with = "color_translate")]
+    #[schemars(schema_with = "schema_color")]
     pub default_color: Color,
     #[serde(default = "dt_focus_color")]
-    #[serde(deserialize_with = "common::color_translate")]
-    #[schemars(schema_with = "common::schema_color")]
+    #[serde(deserialize_with = "color_translate")]
+    #[schemars(schema_with = "schema_color")]
     pub focus_color: Color,
     #[serde(default = "dt_active_color")]
-    #[serde(deserialize_with = "common::color_translate")]
-    #[schemars(schema_with = "common::schema_color")]
+    #[serde(deserialize_with = "color_translate")]
+    #[schemars(schema_with = "schema_color")]
     pub active_color: Color,
     #[serde(default)]
-    #[serde(deserialize_with = "common::option_color_translate")]
-    #[schemars(schema_with = "common::schema_optional_color")]
+    #[serde(deserialize_with = "option_color_translate")]
+    #[schemars(schema_with = "schema_optional_color")]
     pub hover_color: Option<Color>,
 
     #[serde(default)]
